resolved merge conflicts

+++ src/globals.h
@@ -54,5 +54,5 @@ extern std::set<uint16_t> macs;
 
 #ifdef BLECOUNTER
     extern int scanTime;
-    extern std::set<uint32_t> bles; 
+    extern std::set<uint16_t> bles; 
 #endif
